+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # IoC container for Pip.Services in Python Changelog
 
+## <a name="2.1.1"></a> 2.1.1 (2017-04-12)
+
+### Features
+* **container** Now supports IConfigurable, IReferenceable, IUnreferenceable and IOpenable interfaces
+
+### Bug Fixes
+* Container start and stop methods were renamed to open and close
+
 ## <a name="2.1.0"></a> 2.1.0 (2017-04-11)
 
 ### Bug Fixes